set different priorities for different branches
Up to now the default was set for all branches. A priority of 0 however
means an "instant" update leading to a possibly high load on the net and
the update server. The idea of GLUON_PRIORITY was to spread this. So the
values used now are 14 days for stable, 3 days for beta and still 0 for
experimental. This means stable updates come after some time but
eventually after two weeks, beta after at most three days and
experimental every night.