Add global IPv6 prefix via extra_prefixes6
I'm using the (still pending) gluon-alt-esc package [0] here
in Lübeck and for that it'd be nice to have the extra_prefixes6
added.

Another potential user of this value is the
gluon-ebtables-source-filter package. However that one is not
in use in Lübeck by anyone so far as far as I know.

[0]: https://github.com/freifunk-gluon/gluon/pull/1094