Announce fastd public key in respondd/alfred
This disables mesh_vpn.pubkey_privacy and by that publishes the fastd
public key in the mesh via respondd/alfred.

This allows us to ease or even automize the addition of fastd keys for
nodes which were already seen before, for instance via Wifi-Mesh.

The disadvantage is that ISPs could in theory identify which Freifunk
node certain encrypted traffic belongs to via this public key.
However, they could probably already do so by probing mesh nodes
and checking the timings of the encrypted packets.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>